feat(front) add checkout box in cart page
