+++ frontend/src/routes/routes.tsx
@@ -1,13 +1,12 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { UserList } from "../pages/User/UserList";
-import { Products } from "../pages/Product/Products";
 import { Home } from "../pages/Home/Home";
-import { Cart } from "../pages/Cart/Cart"
+import { Checkout } from "../pages/Checkout/Checkout"
 import { DogItems } from "../pages/DogItems/DogItems"; 
 import { ProductList } from "../pages/Product/ProductList";
 import { TopbarLayout } from "../layouts/TopbarLayout";
 import { ProductDetails } from "../pages/Product/ProductDetails";
- 
+
 const AppRoutes = () => {
     return (
         <BrowserRouter>
@@ -15,7 +14,7 @@ const AppRoutes = () => {
                 <Route element={<TopbarLayout />}>
                     <Route path="/" element={<Home />} />
                     <Route path="/product/:id" element={<ProductDetails />} />
-                    <Route path="/cart" element={<Cart/>} />
+                    <Route path="/checkout" element={<Checkout/>} />
                     <Route path="/dog-items" element={<DogItems/>} />
                 </Route>
                 <Route path="/user-list" element={<UserList />} />
